\name{functionrunner-package}
\alias{functionrunner-package}
\alias{functionrunner}
\docType{package}
\title{\packageTitle{functionrunner}}
\description{\packageDescription{functionrunner}}
\section{Package Content}{\packageIndices{functionrunner}}
\author{\packageAuthor{functionrunner}}
\section{Maintainer}{\packageMaintainer{functionrunner}}
\keyword{package}
