% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunction.R
\name{run_function}
\alias{run_function}
\title{Extremely simplified version of tinysnapshot::expect_snapshot_plot}
\usage{
run_function(fn)
}
\arguments{
\item{fn}{function to be run}
}
\description{
Takes a function as an argument and runs that function
}
